requirements.txt
colorlog
exchange-calendars
ib_async
nest-asyncio
pandas
pgvector
psycopg2-binary
pydantic
python-dotenv
pyyaml
sentence-transformers
sqlalchemy
watchdog
xai-sdk
